Update STJ version number

+++ iothub/service/samples/getting started/JobsSample/JobsSample.cs
@@ -12,7 +12,7 @@ namespace Microsoft.Azure.Devices.Samples.JobsSample
     {
         private const string DeviceId = "JobsSample_Device";
         private const string TestTagName = "JobsSample_Tag";
-        private const long TestTagValue = 100;
+        private const int TestTagValue = 100;
 
         private readonly IotHubServiceClient _jobClient;
